% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{cppdoc_index}
\alias{cppdoc_index}
\title{Create package index}
\usage{
cppdoc_index(path = ".", quiet_doxygen = TRUE)
}
\arguments{
\item{path}{Path to the package, default is the current directory.}

\item{quiet_doxygen}{Logical, indicating if doxygen should be run
quietly.  It's pretty verbose so is set to be quiet by default.}
}
\value{
Invisibly, the package index, which is a \code{data.frame} with
class \code{cppdoc_index}.  It is also written into the package
directory at \code{inst/cppdoc/index.rds}.
}
\description{
Build an index for a package.  This requires that the doxygen is
installed and available.
}
