% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{cppdoc_function}
\alias{cppdoc_function}
\alias{cppdoc_define}
\alias{cppdoc_typedef}
\alias{cppdoc_class}
\alias{cppdoc_enum}
\alias{cppdoc_example}
\title{Create markdown output}
\usage{
cppdoc_function(name, args = NULL, package = NULL)

cppdoc_define(name, package = NULL)

cppdoc_typedef(name, package = NULL)

cppdoc_class(name, package = NULL)

cppdoc_enum(name, package = NULL)

cppdoc_example(name, input = TRUE, output = TRUE)
}
\arguments{
\item{name}{The name of the entity to render}

\item{args}{For \code{cppdoc_function}, optionally argument types (not
optional in the case of overloaded functions where this is used
to select the appropriate method).}

\item{package}{Optionally the package name.  This can be \code{NULL}
after running \code{cppdoc_register}, typically within a knitr
document.}

\item{input, output}{Logical, indicating if input or output should
be produced.}
}
\value{
A character vector with markdown-formatted text, typically
for inclusion in a knitr document via the \code{cppdoc} engine
}
\description{
Create markdown output from C++ entities.
}
