% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{cppdoc_register}
\alias{cppdoc_register}
\alias{cppdoc_unregister}
\title{Prepare for use with knitr}
\usage{
cppdoc_register(package)

cppdoc_unregister()
}
\arguments{
\item{package}{Name of the package to use}
}
\value{
Nothing, called for its side effects
}
\description{
Prepare for use with knitr by registering the \code{cppdoc} knitr
engine and optionally setting a default package to locate C++
definitions in.  This operation can be undone with
\code{cppdoc_unregister}.
}
\examples{
cppdoc::cppdoc_register(NULL)
cppdoc::cppdoc_unregister()
}
